CRM_Project/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.8
django-redis==5.3.0
celery==5.3.4
redis==4.5.5
